+++ packages/typedoc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @to-skills/typedoc
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @to-skills/core@1.2.0
+
 ## 1.0.4
 
 ### Patch Changes